+++ packages/api-client/tests/test-granules.js
@@ -373,6 +373,35 @@ test('createGranule calls the callback with the expected object', async (t) => {
   }));
 });
 
+test('replaceGranule calls the callback with the expected object', async (t) => {
+  const body = {
+    granuleId: t.context.granuleId,
+    any: 'object',
+    status: t.context.status,
+  };
+  const expected = {
+    prefix: t.context.testPrefix,
+    payload: {
+      httpMethod: 'PUT',
+      resource: '/{proxy+}',
+      path: `/granules/${t.context.granuleId}`,
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body),
+    },
+    expectedStatusCodes: [200, 201],
+  };
+
+  const callback = (configObject) => {
+    t.deepEqual(configObject, expected);
+  };
+
+  await t.notThrowsAsync(granulesApi.replaceGranule({
+    callback,
+    prefix: t.context.testPrefix,
+    body,
+  }));
+});
+
 test('updateGranule calls the callback with the expected object', async (t) => {
   const body = {
     granuleId: t.context.granuleId,